requirements.txt
django
djangorestframework
mysqlclient
pyserial
python-dateutil
python-dateutil --upgrade
requests
django-cleanup
django_resized